[RCL-170] FIX: read_civis now reads small tables in memory, large tables from file.
The total number of bytes in a textConnection is limited to 2^31-1, the maximum
size of a string in R. This limit means that reading tables
in read_civis straight from a textConnection is limited to tables of ~2gb.

read_civis now checks the size of the file before reading. Files larger than
2^31-1 bytes are downloaded to a temporary file. Files < 2^31-1 bytes are still
read straight from a textConnection.

In addition, the following changes were made:
- If a query returns no results, this is now signalled in read_civis.sql and download_script_results (for download_civis)
- `stop_if_no_output` throws an error if a query produces no output.
- refactored tests for read_civis to add clarity.