Patch fetch_all pagination helper (#243)
* Patch fetch_all pagination helper

The `fetch_all` function is intended to call `fetch_until` until a result of `FALSE` is returned. However, when the API returns no result, that output manifests as a NULL, not a boolean `FALSE` value. This change prevents a coercion error of trying to boolean compare a NULL value in R.

* ADD Bryan Baird as a contributor

* Update DESCRIPTION

---------

Co-authored-by: pcooman <pcooman@civisanalytics.com>