Prerelease Fix 3.1.3 (#260)
* FIX restore escape characters.

This is a follow up to #259 which removed the escape strings. This was
an issue for the last 3.1.2 build according to CRAN but it wasn't clear
to me that it was a blocker for publishing to CRAN. At any rate, this
build now runs with no notes locally.

* BLD ignore github directory.

* fixup! FIX restore escape characters.

* DOC address a couple doc NOTES from window build.

* DOC link plans to future package.

Flagged as a NOTE by CRAN.